requirements.txt
selenium
python-dotenv
reportlab
PyPDF2
pillow
google-generativeai
pandas
langchain-google-genai
browser-use